requirements.txt
torch
transformers
scikit-learn
opencv-python
matplotlib
Pillow
